requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
safetensors>=0.3.0
Pillow>=9.0.0
requests>=2.28.0
peft>=0.6.0
